requirements.txt
Flask==2.2.2
web3>=6.10.0
python-dotenv==0.21.0
Flask-SQLAlchemy==3.0.2
Werkzeug==2.2.2